Rename JsonString to JsonBuilder
