+++ fdbclient/JsonBuilder.cpp
@@ -1,4 +1,4 @@
-#include "JsonString.h"
+#include "JsonBuilder.h"
 #include <iostream>
 
 JsonBuilderObject JsonBuilder::makeMessage(const char *name, const char *description) {
