+++ fdbserver/Status.actor.cpp
@@ -32,7 +32,7 @@
 #include "flow/UnitTest.h"
 #include "QuietDatabase.h"
 #include "RecoveryState.h"
-#include "fdbclient/JsonString.h"
+#include "fdbclient/JsonBuilder.h"
 
 const char* RecoveryStatus::names[] = {
 	"reading_coordinated_state", "locking_coordinated_state", "locking_old_transaction_servers", "reading_transaction_system_state",
